Move tests.MockKibanaClient to its own package (#5273)
