+++ kibana/kibanatest/kibana.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package tests
+package kibanatest
 
 import (
 	"context"
